extract initial assignment
