+++ sha/src/circuit/synthesis.rs
@@ -1,3 +1,5 @@
 mod limb_decomposition;
+mod initial_assignment;
 
 pub use limb_decomposition::*;
+pub use initial_assignment::*;
